move shared markdown functionality into @os/lib/markdown
