+++ src/os/lib/markdown/codemirrorPlugins/codeMonospace.ts
