+++ src/os/lib/markdown/index.ts
@@ -1,0 +1,3 @@
+export { MarkdownInput } from "./markdownInput";
+export { useMarkdownPlugins } from "./useMarkdownPlugins";
+export { theme } from "./codemirrorPlugins/theme";
